# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for finAns personal finance admin",
        "title": "finAns API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/analytics/expenses-calendar": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Expenses calendar aggregation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "day or month",
                        "name": "level",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "year",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "month (1-12), required for level=day",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/health": {
            "get": {
                "description": "Returns service and database connectivity status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HealthResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/import/batches": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Upload a CSV file for import moderation",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ImportBatchWithRowsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/import/batches/active": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Get the active (not yet closed) import batch, if any",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ImportBatchWithRowsResponse"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/import/batches/{id}/accept": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Accept multiple ready moderation rows into transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AcceptedTransactionResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/import/batches/{id}/close": {
            "post": {
                "tags": [
                    "import"
                ],
                "summary": "Close an import batch (discard the moderation session)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/import/rows/{id}": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Inline-edit a moderation row",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Row ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ModerationRowResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/import/rows/{id}/accept": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Accept a single ready moderation row into transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Row ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.AcceptedTransactionResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/planned-expense-categories": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planned-expense-categories"
                ],
                "summary": "List planned expense categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planned-expense-categories"
                ],
                "summary": "Create a planned expense category",
                "parameters": [
                    {
                        "description": "Category",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePlannedExpenseCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/planned-expense-categories/reorder": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "planned-expense-categories"
                ],
                "summary": "Persist a new drag-and-drop order for category cards",
                "parameters": [
                    {
                        "description": "Full ordered list of category IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReorderPlannedExpenseCategoriesRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/planned-expenses": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planned-expenses"
                ],
                "summary": "List planned expenses (active grouped by category, or archived flat)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "active (default) or archived",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planned-expenses"
                ],
                "summary": "Create a planned expense item",
                "parameters": [
                    {
                        "description": "Item",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePlannedExpenseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/planned-expenses/{id}": {
            "delete": {
                "tags": [
                    "planned-expenses"
                ],
                "summary": "Delete a planned expense item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planned-expenses"
                ],
                "summary": "Update a planned expense item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePlannedExpenseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/planned-expenses/{id}/complete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planned-expenses"
                ],
                "summary": "Mark a planned expense item as done and move it to the archive",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/transactions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search title",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AcceptedTransactionResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "specificity": {
                    "type": "string"
                },
                "tag_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.CreatePlannedExpenseCategoryRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.CreatePlannedExpenseRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "cost_kopecks": {
                    "type": "integer"
                },
                "due_date": {
                    "type": "string"
                },
                "new_category": {
                    "$ref": "#/definitions/dto.CreatePlannedExpenseCategoryRequest"
                },
                "priority": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.HealthResponse": {
            "type": "object",
            "properties": {
                "db": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dto.ImportBatchResponse": {
            "type": "object",
            "properties": {
                "closed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "dto.ImportBatchWithRowsResponse": {
            "type": "object",
            "properties": {
                "batch": {
                    "$ref": "#/definitions/dto.ImportBatchResponse"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ModerationRowResponse"
                    }
                }
            }
        },
        "dto.ModerationRowResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "batch_id": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "field_errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "row_number": {
                    "type": "integer"
                },
                "specificity": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tag_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.ReorderPlannedExpenseCategoriesRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dto.UpdatePlannedExpenseRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "cost_kopecks": {
                    "type": "integer"
                },
                "due_date": {
                    "type": "string"
                },
                "new_category": {
                    "$ref": "#/definitions/dto.CreatePlannedExpenseCategoryRequest"
                },
                "priority": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}